Update: /constants.dart --> fixing the baseUrl and the arguments /remote_data_source.dart --> arguments positioning /current_weather_mdeol --> change the model property naming and fromJson /weather_repository_impl.dart -_> argument positioning /current_weather.dart --> fixing the entity property /weather_repository.dart -->  argument positioning /get_current_weather.dart --> argument positioning /page_current_weather.dart -->  change to Stateful widget and bloc init /remote_data_source_test.dart --> fixing test property and argument positioning /weather_repository_impl_test.dart --> code cleanups /weather_bloc_test.dart --> code celeanups

+++ lib/domain/entities/current_weather.dart
@@ -21,11 +21,11 @@ class CurrentWeather extends Equatable {
   final double temperature;
   final double minTemperature;
   final double maxTemperature;
-  final double pressure;
-  final double humidity;
+  final int pressure;
+  final int humidity;
   final double winSpeed;
   final String cloud;
-  final double visibility;
+  final int visibility;
 
   @override
   List<Object?> get props => [
